🚀 RSI 전략과 동일한 구조의 MACD 전략 클래스 구현 완료

+++ trading_bot.py
@@ -697,7 +697,7 @@ RSI: {rsi:.1f}
             
             # 체결 금액 계산
             try:
-                qty = float(execution_qty)
+                qty = int(execution_qty)
                 price = float(int(execution_price)/10000)
                 total_amount = qty * price
             except:
